tests: move actions data fixture to conftest
