+++ tests/streams/actions/test_files_actions.py
@@ -21,76 +21,12 @@ from invenio_rdm_migrator.streams.models.files import (
 from invenio_rdm_migrator.streams.models.records import RDMDraftFile
 
 
-@pytest.fixture(scope="function")
-def bucket_data():
-    """Bucket data."""
-    return {
-        "id": "0e12b4b6-9cc7-46df-9a04-c11c478de211",
-        "created": "2023-06-29T13:00:00",
-        "updated": "2023-06-29T14:00:00",
-        "default_location": 1,
-        "default_storage_class": "S",
-        "size": 1562554,
-        "quota_size": 50000000000,
-        "max_file_size": 50000000000,
-        "locked": False,
-        "deleted": False,
-    }
-
-
-@pytest.fixture(scope="function")
-def ov_data():
-    """Object version data."""
-    return {
-        "version_id": "f8200dc7-55b6-4785-abd0-f3d13b143c98",
-        "created": "2023-06-29T13:00:00",
-        "updated": "2023-06-29T14:00:00",
-        "key": "IMG_3535.jpg",
-        "bucket_id": "0e12b4b6-9cc7-46df-9a04-c11c478de211",
-        "file_id": "e94b243e-9c0c-44df-bd1f-6decc374cf78",
-        "_mimetype": None,
-        "is_head": True,
-    }
-
-
-@pytest.fixture(scope="function")
-def fi_data():
-    """File instance data."""
-    return {
-        "id": "e94b243e-9c0c-44df-bd1f-6decc374cf78",
-        "created": "2023-06-29T13:00:00",
-        "updated": "2023-06-29T14:00:00",
-        "uri": "root://eosmedia.cern.ch//eos/media/zenodo/test/data/e9/4b/243e-9c0c-44df-bd1f-6decc374cf78/data",
-        "storage_class": "S",
-        "size": 1562554,
-        "checksum": "md5:3cc016be06f2be46d3a438db23c40bf3",
-        "readable": True,
-        "writable": False,
-        "last_check_at": None,
-        "last_check": True,
-    }
-
-
-@pytest.fixture(scope="function")
-def fr_data(ov_data):
-    """File record data."""
-    return {
-        "id": None,
-        "json": {},
-        "created": ov_data["created"],
-        "updated": ov_data["updated"],
-        "version_id": 1,
-        "key": ov_data["key"],
-        "record_id": None,
-        "object_version_id": ov_data["version_id"],
-    }
-
-
-# TODO: add to state the bucket
-def test_upload_file_action(buckets_state, bucket_data, fi_data, ov_data, fr_data):
+def test_upload_file_action(
+    buckets_state, bucket_data_w_size, fi_data, ov_data, fr_data
+):
     data = dict(
         tx_id=1,
-        bucket=bucket_data,
+        bucket=bucket_data_w_size,
         object_version=ov_data,
         file_instance=fi_data,
         file_record=fr_data,
